pyproject.toml
[project]
name = "playwright-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.15.0",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
